bump avalanchego to master
